feat: used Modal component
