+++ app/components/Modal.tsx
@@ -1,3 +1,5 @@
+import Image from 'next/image'
+
 export function Modal({
   closeModal,
   children,
@@ -6,27 +8,19 @@ export function Modal({
   children: React.ReactNode
 }) {
   return (
-    <div className="fixed inset-0 z-40 flex items-center justify-center overflow-y-auto bg-black/20">
-      <div className={`relative my-6 max-h-[90vh] shadow-medium`}>
+    <div className="fixed inset-0 z-40 flex items-center justify-center bg-black bg-opacity-50">
+      <div className="relative max-h-[90vh] w-full max-w-3xl overflow-y-auto rounded-2xl bg-white px-16 pb-16 pt-[52px]">
         <button
+          className="absolute right-4 top-4 text-gray-500 hover:text-gray-700"
           onClick={closeModal}
-          className={`absolute right-5 top-5 h-11 w-11 items-center justify-center p-3 text-black`}
         >
-          <svg
-            width="16"
-            height="16"
-            viewBox="0 0 16 16"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <path
-              d="M2 2L8 8M14 14L8 8M8 8L14 2L2 14"
-              stroke="currentColor"
-              strokeWidth="3"
-              strokeLinecap="round"
-              strokeLinejoin="round"
-            />
-          </svg>
+          <Image
+            src="/icons/close.svg"
+            alt="Close"
+            width={24}
+            height={24}
+            priority={true}
+          />
         </button>
         {children}
       </div>
